edit methods in UserController

+++ app/Http/Controllers/Dashboard/UserController.php
@@ -12,6 +12,11 @@ use Yajra\DataTables\DataTables;
 
 class UserController extends Controller
 {
+    protected $user;
+    public function __construct(User $user)
+    {
+        $this->user = $user;
+    }
     public function index()
     {
         return view('dashboard.users.index');
@@ -23,45 +28,32 @@ class UserController extends Controller
 
     public function getUsersDatatable()
     {
-
-        $data = User::select('*');
-        return DataTables::of($data)->addIndexColumn()
-            ->addColumn('action',function ($row){
-                return $btn = '
-                <a href="'.Route('dashboard.users.edit',$row->id).'" class="edit btn btn-success btn-sm"><i class="fa fa-edit"></i><a/>
-                <a id="deleteBtn" data-id="'.$row->id.'" class="edit btn btn-danger btn-sm" data-toggle="modal"
-                data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
+        if (auth()->user()->can('viewAny', $this->user)) {
+            $data = User::select('*');
+        }else{
+            $data = User::where('id' , auth()->user()->id);
+        }
+        return   Datatables::of($data)
+            ->addIndexColumn()
+            ->addColumn('action', function ($row) {
+                $btn = '';
+                if (auth()->user()->can('update', $row)) {
+                    $btn .= '<a href="' . Route('dashboard.users.edit', $row->id) . '"  class="edit btn btn-success btn-sm" ><i class="fa fa-edit"></i></a>';
+                }
+                if (auth()->user()->can('delete', $row)) {
+                    $btn .= '<a id="deleteBtn" data-id="' . $row->id . '" class="edit btn btn-danger btn-sm"  data-toggle="modal" data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
+                }
+                return $btn;
             })
             ->addColumn('status', function ($row) {
                 return $row->status == null ? __('words.not activated') : __('words.' . $row->status);
             })
-//        if (auth()->user()->can('viewAny', $this->user)) {
-//            $data = User::select('*');
-//        }else{
-//            $data = User::where('id' , auth()->user()->id);
-//        }
-//        return Datatables::of($data)
-//            ->addIndexColumn()
-//            ->addColumn('action', function ($row) {
-//                $btn = '';
-//                if (auth()->user()->can('update', $row)) {
-//                    $btn .= '<a href="' . Route('dashboard.users.edit', $row->id) . '"  class="edit btn btn-success btn-sm" ><i class="fa fa-edit"></i></a>';
-//                }
-//                if (auth()->user()->can('delete', $row)) {
-//                    $btn .= '
-//
-//                        <a id="deleteBtn" data-id="' . $row->id . '" class="edit btn btn-danger btn-sm"  data-toggle="modal" data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
-//                }
-//                return $btn;
-//            })
-//            ->addColumn('status', function ($row) {
-//                return $row->status == null ? __('words.not activated') : __('words.' . $row->status);
-//            })
             ->rawColumns(['action', 'status'])
             ->make(true);
     }
     public function store(StoreUserRequest $request)
     {
+        $this->authorize('update', $this->user);
         User::create([
             'name' => $request->name,
             'status' => $request->status,
@@ -84,6 +76,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
         return view('dashboard.users.edit' ,compact('user'));
     }
 
@@ -92,10 +85,11 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
-//        $request->validate([
-//            'name' => 'required|string|max:255',
-//            'email' => 'nullable|string|email|max:255,',
-//        ]);
+        $this->authorize('update', $user);
+        $request->validate([
+            'name' => 'required|string|max:255',
+            'email' => 'nullable|string|email|max:255,',
+        ]);
         $user->update($request->all());
         return redirect()->route('dashboard.users.index');
     }
@@ -109,6 +103,7 @@ class UserController extends Controller
     }
     public function delete(Request $request)
     {
+        $this->authorize('delete', $this->user);
         if (is_numeric($request->id))
         {
             User::where('id',$request->id)->delete();
